refactor: enhance path validation diagnostics and refine build process
- Update path_sanity_strerror to accept a path argument.
- Modify prepare_config_path to automatically resolve relative paths
into absolute paths using the current working directory.
- Fix the multi-stage PGO build and ThinLTO generation in the Makefile.
Isolate Clang optimization records into a dedicated build/opt_records
directory.
- Remove hardcoded escaped quotes for DATADIR and SYSCONFDIR in the
Makefile and introduce a standard STR() macro stringification approach
in help.c and stowignore.c
- Add missing documentation for ignore commands to the help.txt
resource.
- Resolve pointer arithmetic type-casting warnings in path.c, remove the
unused g_signal_temp_count variable in signal.c, and adapt unit tests to
match the new path sanity string outputs.

+++ src/utils/path.c
@@ -47,7 +47,7 @@ void normalize_path(char *path)
         r++;
     }
     *w = '\0';
-    size_t len = w - path;
+    size_t len = (size_t)(w - path);
     if (len > 1 && path[len - 1] == '/') {
         path[len - 1] = '\0';
     }
@@ -87,7 +87,7 @@ void collapse_path(char *path)
         if (len == 2 && start[0] == '.' && start[1] == '.') {
             if (depth > 0) {
                 size_t prev = stack[depth - 1];
-                if (!is_abs && (size_t)(w - path - prev) == 2 && path[prev] == '.' &&
+                if (!is_abs && (size_t)(w - (size_t)path - prev) == 2 && path[prev] == '.' &&
                     path[prev + 1] == '.') {
                     depth++;
                     if (w > path && *(w - 1) != '/') {
